Use DiscordService to fetch user by username
